Use ICustomMarshaler for missing UnmanagedType.LPUTF8Str in .NET Standard 2.0 or lower (#212)

+++ FFmpeg.AutoGen.Examples.ResamplingAudio/Program.cs
@@ -7,8 +7,8 @@ namespace FFmpeg.AutoGen.Examples.ResamplingAudio
     {
         static int Main(string[] args)
         {
-            int sourceChannelLayout = ffmpeg.AV_CH_LAYOUT_STEREO;
-            int destinationChannelLayout = ffmpeg.AV_CH_LAYOUT_SURROUND;
+            long sourceChannelLayout = (long)ffmpeg.AV_CH_LAYOUT_STEREO;
+            long destinationChannelLayout = (long)ffmpeg.AV_CH_LAYOUT_SURROUND;
             int sourceSampleRate = 48000;
             int destinationSampleRate = 44100;
             byte** sourceData = null;
